% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{empregos}
\alias{empregos}
\title{Empregos no setor automotivo desde 1983}
\format{
Um data frame com 864 linhas e 3 variaveis:
\describe{
  \item{mes}{Mes}
  \item{tipo}{Setor}
  \item{valor}{Empregos}
}
}
\source{
\url{http://www.anfavea.com.br/estatisticas.html}
}
\usage{
empregos
}
\description{
Uma tabela contendo os empregos mensais no setor automotivo do Brasil
 desde janeiro de 1983.
}
\keyword{datasets}
