% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exportacoes}
\alias{exportacoes}
\title{Exportacao de veiculos e maquinas desde 1985}
\format{Um data frame com 816 linhas e 3 variaveis:
\describe{
  \item{mes}{Mes}
  \item{tipo}{Setor}
  \item{valor}{Exportacoes, em mil US$}
}}
\source{
\url{http://www.anfavea.com.br/estatisticas.html}
}
\usage{
exportacoes
}
\description{
Uma tabela contendo as exportacao mensais de veiculos e maquinas no
 Brasil desde janeiro de 1985.
}
\keyword{datasets}
