% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maquinas}
\alias{maquinas}
\title{Vendas, producao e exportacao de maquinas desde 1960}
\format{Um data frame com 24780 linhas e 4 variaveis:
\describe{
  \item{mes}{Mes}
  \item{maquina}{Tipo de maquina}
  \item{variavel}{Variavel}
  \item{valor}{Valor}
}}
\source{
\url{http://www.anfavea.com.br/estatisticas.html}
}
\usage{
maquinas
}
\description{
Uma tabela contendo a quantidade de vendas, producao e exportacao mensais de
 maquinas no Brasil desde janeiro de 1960.
}
\keyword{datasets}
