% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{veiculos}
\alias{veiculos}
\title{Vendas, producao e exportacao de veiculos desde 1957}
\format{
Um data frame com 18600 linhas e 4 variaveis:
\describe{
  \item{mes}{Mes}
  \item{veiculo}{Tipo de veiculo}
  \item{variavel}{Variavel}
  \item{valor}{Valor}
}
}
\source{
\url{http://www.anfavea.com.br/estatisticas.html}
}
\usage{
veiculos
}
\description{
Uma tabela contendo a quantidade de licenciamentos, producao e exportacao
 mensais de veiculos no Brasil desde janeiro de 1957.
}
\keyword{datasets}
