% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{empregos}
\alias{empregos}
\title{Empregos no setor automotivo desde 1983}
\format{Um data frame com 432 linhas e 4 variaveis:
\describe{
  \item{mes}{Mes}
  \item{total}{Total de empregos em ambos setores}
  \item{veiculos}{Empregos para producao de veiculos}
  \item{maquinas}{Empregos para producao de maquinas}
}}
\source{
\url{http://www.anfavea.com.br/estatisticas.html}
}
\usage{
empregos
}
\description{
Uma tabela contendo os empregos mensais no setor automotivo do Brasil
 desde janeiro de 1983.
}
\keyword{datasets}
